ui config for outside app
